[modify] 프로필 데이터 변수 수정

+++ app/src/main/java/org/personal/coupleapp/data/ProfileData.kt
@@ -1,3 +1,3 @@
 package org.personal.coupleapp.data
 
-data class ProfileData(val id: Int, var profile_image: Any?, val name: String, val state_message: String?, val birthday: Int, val sex: String)
+data class ProfileData(val id: Int, var profile_image: Any?, val name: String, val state_message: String?, val birthday: Long, val sex: String)
